+++ src/drawer.canvasentry.js
@@ -179,12 +179,20 @@ export default class CanvasEntry {
      * @param {number} y Y start position
      * @param {number} width Width of the rectangle
      * @param {number} height Height of the rectangle
+     * @param {number} radius Radius of the rectangle
      */
-    fillRects(x, y, width, height) {
-        this.fillRectToContext(this.waveCtx, x, y, width, height);
+    fillRects(x, y, width, height, radius) {
+        this.fillRectToContext(this.waveCtx, x, y, width, height, radius);
 
         if (this.hasProgressCanvas) {
-            this.fillRectToContext(this.progressCtx, x, y, width, height);
+            this.fillRectToContext(
+                this.progressCtx,
+                x,
+                y,
+                width,
+                height,
+                radius
+            );
         }
     }
 
@@ -197,12 +205,61 @@ export default class CanvasEntry {
      * @param {number} y Y start position
      * @param {number} width Width of the rectangle
      * @param {number} height Height of the rectangle
+     * @param {number} radius Radius of the rectangle
      */
-    fillRectToContext(ctx, x, y, width, height) {
+    fillRectToContext(ctx, x, y, width, height, radius) {
         if (!ctx) {
             return;
         }
-        ctx.fillRect(x, y, width, height);
+
+        if (radius) {
+            this.drawRoundedRect(ctx, x, y, width, height, radius);
+        } else {
+            ctx.fillRect(x, y, width, height);
+        }
+    }
+
+    /**
+     * Draw a rounded rectangle on Canvas
+     *
+     * @private
+     * @param {CanvasRenderingContext2D} ctx Canvas context
+     * @param {number} x X-position of the rectangle
+     * @param {number} y Y-position of the rectangle
+     * @param {number} width Width of the rectangle
+     * @param {number} height Height of the rectangle
+     * @param {number} radius Radius of the rectangle
+     *
+     * @return {void}
+     * @example drawRoundedRect(ctx, 50, 50, 5, 10, 3)
+     */
+    drawRoundedRect(ctx, x, y, width, height, radius) {
+        if (height === 0) {
+            return;
+        }
+        // peaks are float values from -1 to 1. Use absolute height values in
+        // order to correctly calculate rounded rectangle coordinates
+        if (height < 0) {
+            height *= -1;
+            y -= height;
+        }
+        ctx.beginPath();
+        ctx.moveTo(x + radius, y);
+        ctx.lineTo(x + width - radius, y);
+        ctx.quadraticCurveTo(x + width, y, x + width, y + radius);
+        ctx.lineTo(x + width, y + height - radius);
+        ctx.quadraticCurveTo(
+            x + width,
+            y + height,
+            x + width - radius,
+            y + height
+        );
+        ctx.lineTo(x + radius, y + height);
+        ctx.quadraticCurveTo(x, y + height, x, y + height - radius);
+        ctx.lineTo(x, y + radius);
+        ctx.quadraticCurveTo(x, y, x + radius, y);
+        ctx.closePath();
+        ctx.fill();
     }
 
     /**