+++ CHANGES.md
@@ -1,6 +1,47 @@
 wavesurfer.js changelog
 =======================
 
+3.2.0 (unreleased)
+------------------
+
+- New `MediaElementWebAudio` backend (#1767):
+  - Allows you to use Web Audio API with big audio files, loading audio
+    like with MediaElement backend (HTML5 audio tag), so you can use the
+    same methods of MediaElement backend for loading and playback. This way,
+    the audio resource is not loaded entirely from server, but in ranges,
+    allowing you to use WebAudio features, like filters, on audio files with
+    a long duration. You can also supply peaks data, so the entire audio file
+    does not have to be decoded.
+    For example:
+    ` wavesurfer.load(url | HTMLMediaElement, peaks, preload, duration);
+      wavesurfer.play();
+      wavesurfer.setFilter(customFilter);
+    `
+- Add `barRadius` option to create waveforms with rounded bars (#953)
+- Throw error when the url parameter supplied to `wavesurfer.load()`
+  is empty (#1773, #1775)
+- Specify non-minified wavesurfer.js in `main` entry of `package.json` (#1759)
+- Add `dblclick` event listener to wavesurfer wrapper (#1764)
+- Cursor plugin: flip position of time text to left of the cursor where needed
+  to improve readability (#1776)
+- Regions plugin: change region end handler position (#1762)
+
+3.1.0 (26.09.2019)
+------------------
+
+- Add `autoCenter` and `autoCenterRate` options (#1699)
+- Make sure `isReady` is true before firing the `ready` event (#1749)
+- Improve fetch error messages (#1748)
+- Use `MediaElement` backend for browsers that don't support WebAudio (#1739)
+- Regions plugin:
+  - Use `isResizing` and `isDragging` to filter events in
+    region-updated listener (#1716)
+  - Fix `playLoop` and `loop` option for clips with duration <15s (#1626)
+- Spectrogram plugin: fix variable name in click handler (#1742)
+- Minimap plugin: fix left/width calculations for regions on retina/4k
+  screens (#1743)
+- New example: video-annotation (#1726)
+
 3.0.0 (11.07.2019)
 ------------------